+++ src/utils/asyncUtils.ts
@@ -2,7 +2,7 @@ import { RedisClientType } from '@redis/client';
 import { Api, Context, InlineKeyboard } from 'grammy';
 import { ChatFromGetChat, ChatMember } from 'grammy/types';
 
-import ListsNames from '@enums/listsNames';
+import ListsNames from '@data/listsNames';
 
 import ignoreListMessages from '@locale/ignoreListMessages';
 import keyboardMessages from '@locale/keyboardMessages';